use correct cast (is bitmask for < 32bits so casting to Uint32 is okay)

+++ storage/ndb/src/ndbapi/NdbEventOperationImpl.hpp
@@ -649,7 +649,7 @@ private:
   NdbMutex *m_add_drop_mutex;
 
   inline Gci_container* find_bucket(Uint64 gci){
-    Uint32 pos = (gci & ACTIVE_GCI_MASK);
+    Uint32 pos = (Uint32)(gci & ACTIVE_GCI_MASK);
     Gci_container *bucket= ((Gci_container*)(m_active_gci.getBase())) + pos;
     if(likely(gci == bucket->m_gci))
       return bucket;
